requirements.txt
coinbase
Flask
python-bittrex